# openapi.json
{"openapi":"3.1.0","info":{"title":"CareerPilot AI","description":"AI Resume & Interview Assistant","version":"1.0.0"},"paths":{"/api/health":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/resume/upload":{"post":{"tags":["Resume"],"summary":"Upload Resume","operationId":"upload_resume_api_resume_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_api_resume_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/job/match":{"post":{"tags":["Job Matching"],"summary":"Job Match","operationId":"job_match_api_job_match_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobMatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/interview/start":{"post":{"tags":["Interview"],"summary":"Start Interview","operationId":"start_interview_api_interview_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewStartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/interview/evaluate":{"post":{"tags":["Interview"],"summary":"Evaluate Interview","operationId":"evaluate_interview_api_interview_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterviewEvaluateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ats/analyze":{"post":{"tags":["ATS"],"summary":"Analyze Resume For Job","operationId":"analyze_resume_for_job_api_ats_analyze_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_resume_for_job_api_ats_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_analyze_resume_for_job_api_ats_analyze_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"job_title":{"type":"string","title":"Job Title"},"job_description":{"type":"string","title":"Job Description","default":""},"required_keywords":{"type":"string","title":"Required Keywords","default":""},"preferred_keywords":{"type":"string","title":"Preferred Keywords","default":""},"education":{"type":"string","title":"Education","default":""},"experience_min":{"type":"integer","title":"Experience Min","default":0},"experience_max":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Experience Max"}},"type":"object","required":["file","job_title"],"title":"Body_analyze_resume_for_job_api_ats_analyze_post"},"Body_upload_resume_api_resume_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_resume_api_resume_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InterviewEvaluateRequest":{"properties":{"question":{"type":"string","minLength":3,"title":"Question"},"answer":{"type":"string","minLength":1,"title":"Answer"}},"type":"object","required":["question","answer"],"title":"InterviewEvaluateRequest"},"InterviewStartRequest":{"properties":{"resume_text":{"type":"string","minLength":10,"title":"Resume Text"},"role":{"type":"string","title":"Role","default":"Software Engineer"},"job_description":{"type":"string","title":"Job Description","default":""},"required_keywords":{"items":{"type":"string"},"type":"array","title":"Required Keywords","default":[]},"preferred_keywords":{"items":{"type":"string"},"type":"array","title":"Preferred Keywords","default":[]}},"type":"object","required":["resume_text"],"title":"InterviewStartRequest"},"JobMatchRequest":{"properties":{"resume_text":{"type":"string","minLength":10,"title":"Resume Text"},"job":{"$ref":"#/components/schemas/JobRequirement"}},"type":"object","required":["resume_text","job"],"title":"JobMatchRequest"},"JobRequirement":{"properties":{"job_title":{"type":"string","minLength":2,"title":"Job Title"},"required_keywords":{"items":{"type":"string"},"type":"array","title":"Required Keywords"},"preferred_keywords":{"items":{"type":"string"},"type":"array","title":"Preferred Keywords"},"education":{"items":{"type":"string"},"type":"array","title":"Education"},"experience_min":{"type":"integer","minimum":0.0,"title":"Experience Min","default":0},"experience_max":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Experience Max"}},"type":"object","required":["job_title"],"title":"JobRequirement"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}